cli/operator: move leaf builders from node.go to setup.go
Pure move: the 16 helper functions below the StartNodeCmd closure relocate
verbatim to a new setup.go (warnIfSSVAPIAddressUnset, ensureNoMissingKeys,
privateKeyFromKeystore, validateConfig, setupBadgerDB/setupPebbleDB,
applyMigrations, setupOperatorDataStore, ensureOperatorPrivateKey/PubKey,
setupSSVNetwork, setupP2P, decideNetworkKeyProtectors,
probeRemoteNetworkKeyProtector, syncContractEvents, initSlotPruning), leaving
node.go as the command + Run closure. Imports are split accordingly. No behavior
change - the helper bodies and the closure are byte-identical to before.

First structural step toward extracting the Run closure into a testable
run()/assemble() (later commits); shrinking node.go makes those moves reviewable.
